% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivpower-package.R
\docType{package}
\name{bivpower-package}
\alias{bivpower}
\alias{bivpower-package}
\title{Study of Statistical Power of Bivariate Point Pattern Data}
\description{
Accompanion package to an academic paper.
}
\details{
See details in the paper
T Rajala, S Olhede, D.J Murrell (2019): When do we have the power to detect biological interactions in spatial point patterns? *Journal of Ecology*

https://doi.org/10.1111/1365-2745.13080

The package has vignettes.
}
