% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{double_integral_ball}
\alias{double_integral_ball}
\title{Double integral over ball inclusion}
\usage{
double_integral_ball(W, R, r = 0)
}
\arguments{
\item{W}{rectangular owin window object}

\item{R}{radius of the b(o,R)}

\item{r}{optional lower bound}
}
\description{
2D rectangular window.
}
\details{
Evaluate int_W int_W 1_B(x-y)dxdy where B=b(o,r)
}
