% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K12_doublesum.R
\name{g12_doublesum}
\alias{g12_doublesum}
\title{Doublesum in g12 estimation
The double sum of kernels in g12 estimator. No edge correction or anything fancy.}
\usage{
g12_doublesum(xy, mark, r, bw, kernel = 0)
}
\arguments{
\item{xy}{x,y coordineate matrix}

\item{mark}{1/2 mark vector}

\item{r}{ranges}

\item{bw}{bandwidth}

\item{kernel}{kernel, 0:epa 1:box}
}
\description{
Doublesum in g12 estimation
The double sum of kernels in g12 estimator. No edge correction or anything fancy.
}
\details{
unnormalised. Scale with iso_set_cov and pi*r^2 and lambda^2
}
