% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{iso_set_cov}
\alias{iso_set_cov}
\title{Isotropised Set Covariance}
\usage{
iso_set_cov(W, R)
}
\arguments{
\item{W}{owin}

\item{R}{range}
}
\description{
Isotropised Set Covariance
}
\details{
Let gamma_W(x) = |W cap W_x|.

Compute E gamma_W(Ru) over u distributed as unif(unit sphere)
}
