% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-biv-model1.R
\name{sim_biv_model}
\alias{sim_biv_model}
\title{Simulate Simple Bivariate Interaction Model}
\usage{
sim_biv_model(n, theta, sigma, W, free = FALSE, ..., x1 = NULL, asppp = FALSE)
}
\arguments{
\item{n}{vector of length 2, target point counts}

\item{theta}{interaction parameter, >= -1}

\item{sigma}{scale parameter of interaction kernel}

\item{W}{observation window, spatstat owin-object}

\item{free}{if TRUE, sample point counts as they were Poisson}

\item{...}{passed on to rcox-function}

\item{x1}{If a cbind(x,y)-matrix is given, use these as species 1.}

\item{asppp}{If FALSE (default) return a matrix, else a ppp object}
}
\value{
matrix cbind(x,y,m) with columns for x-coordinate, y-coordinate and 1/2 type mark. if asppp=TRUE, returns a spatstat ppp-object.
}
\description{
Species 1 from Poisson, species 2 from product-shot-noise given species 1.
}
