% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{triple_integral_annulus_annulus_mc}
\alias{triple_integral_annulus_annulus_mc}
\title{Triple integral over two ball inclusion of differences}
\usage{
triple_integral_annulus_annulus_mc(W, R, h, n = 5000, pairs = FALSE)
}
\arguments{
\item{W}{owin}

\item{R}{radius of ball b(o,R)}

\item{h}{annulus halfwidth}

\item{n}{sample size}

\item{pairs}{see details}
}
\description{
Use MC sampling
}
\details{
Let B=b(o,r+h)-b(o,r-h). Evaluate int_W int_W int_W 1_B(x-y)1_B(x-z)dxdydz

If R is a vector and pairs = TRUE, returns a matrix where the integration is with respect to all pairs Bi and Bj of R[i] and R[j].
}
